Refactor IPresentation
