+++ src/ShapeCrawler/Presentations/IFooter.cs
@@ -28,9 +28,9 @@ public interface IFooter
 
 internal sealed class Footer : IFooter
 {
-    private readonly PresentationCore presentationCore;
+    private readonly Presentation presentationCore;
 
-    internal Footer(PresentationCore presentationCore)
+    internal Footer(Presentation presentationCore)
     {
         this.presentationCore = presentationCore;
     }
