+++ src/ShapeCrawler/Presentations/Presentation.cs
@@ -1,95 +1,169 @@
-﻿using System.IO;
-using System.Reflection;
-using ShapeCrawler.Presentations;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using DocumentFormat.OpenXml;
+using DocumentFormat.OpenXml.Packaging;
+using DocumentFormat.OpenXml.Validation;
+using ShapeCrawler.Exceptions;
 using ShapeCrawler.Shared;
+using A = DocumentFormat.OpenXml.Drawing;
 
-#pragma warning disable IDE0130
-namespace ShapeCrawler;
-#pragma warning restore IDE0130
+#if NETSTANDARD2_0
+using ShapeCrawler.Extensions;
+#endif
 
-/// <inheritdoc cref="IPresentation"/>
+namespace ShapeCrawler.Presentations;
+
 public sealed class Presentation : IPresentation
 {
-    private IValidateable validateable;
+    private readonly PresentationDocument presDocument;
+    private readonly SlideSize slideSize;
 
-    /// <summary>
-    ///     Opens existing presentation from specified stream.
-    /// </summary>
     public Presentation(Stream stream)
     {
-        this.validateable = new StreamPresentation(stream);
+        this.presDocument = PresentationDocument.Open(stream, true);
+        var sdkMasterParts = this.presDocument.PresentationPart!.SlideMasterParts;
+        this.SlideMasters = new SlideMasterCollection(sdkMasterParts);
+        this.Sections = new Sections(this.presDocument);
+        this.Slides = new SlideCollection(this.presDocument.PresentationPart);
+        this.Footer = new Footer(this);
+        this.slideSize = new SlideSize(this.presDocument.PresentationPart!.Presentation.SlideSize!);
+        this.FileProperties = new(this.presDocument.CoreFilePropertiesPart!);
     }
-
-    /// <summary>
-    ///     Creates a new presentation.
-    /// </summary>
+    
     public Presentation()
     {
-        var assets = new Assets(Assembly.GetExecutingAssembly());
-        var stream = assets.StreamOf("new-presentation.pptx");
-        this.validateable = new StreamPresentation(stream);
-        this.validateable.Metadata.Modified =
-            this.validateable.Metadata.Created = SCSettings.TimeProvider.UtcNow;
+        
     }
 
-    /// <inheritdoc />
-    public ISlideCollection Slides => this.validateable.Slides;
+    public ISlideCollection Slides { get; }
 
-    /// <inheritdoc />
-    public decimal SlideWidth
-    {
-        get => this.validateable.SlideWidth; 
-        set => this.validateable.SlideWidth = value;
-    }
-    
-    /// <inheritdoc />
     public decimal SlideHeight
     {
-        get => this.validateable.SlideHeight;
-        set => this.validateable.SlideHeight = value;
+        get => this.slideSize.Height();
+        set => this.slideSize.UpdateHeight(value);
     }
-    
-    /// <inheritdoc />
-    public ISlideMasterCollection SlideMasters => this.validateable.SlideMasters;
-    
-    /// <inheritdoc />
-    public ISections Sections => this.validateable.Sections;
 
-    /// <inheritdoc />
-    public IFooter Footer => this.validateable.Footer;
-    
-    /// <inheritdoc />
-    public IPresentationMetadata Metadata => this.validateable.Metadata;
-
-    /// <inheritdoc />
-    public void Save() => this.validateable.Save();
-    
-    /// <inheritdoc />
-    public ISlide Slide(int number) => this.validateable.Slide(number);
-    
-    /// <summary>
-    ///     Gets a presentation byte array.
-    /// </summary>
-    public byte[] AsByteArray() => this.validateable.AsByteArray();
-
-    /// <inheritdoc />
-    public void SaveAs(string path)
+    public decimal SlideWidth
     {
-        this.validateable.CopyTo(path);
-        this.validateable = new PathPresentation(path);
+        get => this.slideSize.Width();
+        set => this.slideSize.UpdateWidth(value);
     }
 
-    /// <inheritdoc />
-    public void SaveAs(Stream stream)
+    public ISlideMasterCollection SlideMasters { get; }
+
+    public ISections Sections { get; }
+
+    public IFooter Footer { get; }
+    public IPresentationMetadata Metadata { get; }
+    public ISlide Slide(int number)
     {
-        this.validateable.CopyTo(stream);
-        this.validateable = new StreamPresentation(stream);
+        throw new System.NotImplementedException();
+    }
+
+    public void Save()
+    {
+        throw new System.NotImplementedException();
+    }
+
+    internal FileProperties FileProperties { get; }
+    
+    internal void Copy(Stream stream)
+    {
+        this.FileProperties.Modified = SCSettings.TimeProvider.UtcNow;
+        this.presDocument.Clone(stream);
     }
     
-    /// <summary>
-    ///     Gets Slide Master by number.
-    /// </summary>
-    public ISlideMaster SlideMaster(int number) => this.SlideMasters[number - 1];
+    internal void Copy(string path)
+    {
+        this.FileProperties.Modified = SCSettings.TimeProvider.UtcNow;
+        var cloned = this.presDocument.Clone(path);
+        cloned.Dispose();
+    }
 
-    internal void Validate() => this.validateable.Validate();
+    internal void Validate()
+    {
+        var nonCriticalErrors = new List<string>
+        {
+                "The element has unexpected child element 'http://schemas.openxmlformats.org/drawingml/2006/chart:showDLblsOverMax'.",
+                "The element has invalid child element 'http://schemas.microsoft.com/office/drawing/2017/03/chart:dataDisplayOptions16'. List of possible elements expected: <http://schemas.microsoft.com/office/drawing/2017/03/chart:dispNaAsBlank>.",
+                "The 'uri' attribute is not declared.",
+                "The 'mod' attribute is not declared.",
+                "The 'mod' attribute is not declared.",
+                "The element has unexpected child element 'http://schemas.openxmlformats.org/drawingml/2006/main:noFill'."
+        };
+        var sdkErrors = new OpenXmlValidator(FileFormatVersions.Microsoft365).Validate(this.presDocument);
+        sdkErrors = sdkErrors.Where(errorInfo => !nonCriticalErrors.Contains(errorInfo.Description));
+        sdkErrors = [.. sdkErrors.DistinctBy(errorInfo => new { errorInfo.Description, errorInfo.Path?.XPath })];
+
+        if (sdkErrors.Any())
+        {
+            throw new SCException("Presentation is invalid.");
+        }
+        
+        var errors = ValidateATableRows(this.presDocument);
+        errors = errors.Concat(ValidateASolidFill(this.presDocument));
+        if (errors.Any())
+        {
+            throw new SCException("Presentation is invalid.");
+        }
+    }
+
+    private static IEnumerable<string> ValidateATableRows(PresentationDocument sdkPres)
+    {
+        var aTableRows = sdkPres.PresentationPart!.SlideParts
+            .SelectMany(slidePart => slidePart.Slide.Descendants<A.TableRow>());
+
+        foreach (var aTableRow in aTableRows)
+        {
+            var aExtLst = aTableRow.GetFirstChild<A.ExtensionList>();
+            if (aExtLst != null)
+            {
+                var lastTableCellIndex = -1;
+                var extListIndex = -1;
+
+                // Find indices of last TableCell and ExtensionList
+                for (int i = 0; i < aTableRow.ChildElements.Count; i++)
+                {
+                    var element = aTableRow.ChildElements[i];
+                    if (element is A.TableCell)
+                    {
+                        lastTableCellIndex = i;
+                    }
+                    else if (element is A.ExtensionList)
+                    {
+                        extListIndex = i;
+                    }
+                }
+
+                // If ExtensionList appears before the last TableCell, yield the error
+                if (extListIndex < lastTableCellIndex)
+                {
+                    yield return "Invalid table row structure: ExtensionList element must appear after all TableCell elements in a TableRow";
+                }
+            }
+        }
+    }
+    
+    private static IEnumerable<string> ValidateASolidFill(PresentationDocument sdkPres)
+    {
+        var aText = sdkPres.PresentationPart!.SlideParts
+            .SelectMany(slidePart => slidePart.Slide.Descendants<A.Text>());
+        aText =
+        [
+            .. aText,
+            .. sdkPres.PresentationPart!.SlideMasterParts
+                .SelectMany(slidePart => slidePart.SlideMaster.Descendants<A.Text>()),
+        ];
+
+        foreach (var text in aText)
+        {
+            var runProperties = text.Parent!.GetFirstChild<A.RunProperties>();
+            if ((runProperties?.Descendants<A.SolidFill>().Any() ?? false) 
+                && runProperties.ChildElements.Take(2).All(x => x is not A.SolidFill))
+            {
+                yield return "Invalid solid fill structure: SolidFill element must be index 0";
+            }
+        }
+    }
 }
